change structure of the project
