+++ src/app/footer/footer.css
@@ -1,0 +1,19 @@
+.footer {
+    margin: 0;
+    padding: 20px 0 20px 100px;
+    text-align: left;
+}
+
+.footer span {
+    color: rgb(102, 102, 102);
+    padding: 20px;
+    width: 25%;
+    border-bottom: 1px solid rgb(255, 255, 255);
+    font-size: 18px;
+}
+
+.legalFooter {
+    color: rgb(102, 102, 102);
+    margin: 0;
+    padding: 20px;
+}
