+++ src/app/header/header.css
@@ -1,0 +1,6 @@
+.logo {
+    background-color: rgb(248, 248, 248);
+    margin: 0;
+    border: rgb(20, 20, 20);
+    text-align: left;
+}
