+++ src/app/header/searchField/search.css
@@ -1,0 +1,19 @@
+.searchTitle {
+    color: rgb(102, 102, 102);
+    margin: 0 0 0 120px;
+    padding-top: 20px;   
+    text-align: left;
+    font-size: 1.5em; 
+}
+
+.searchField {
+    margin: 20px 20px 20px 0;
+    padding: 10px 16px;
+    height: 46px;
+    line-height: 1.33;
+    width: 80%;
+    border: 1px solid rgb(180, 180, 180);
+    border-radius: 2px;
+    box-shadow: inset 0 1px 1px rgba(0,0,0,.075);
+    font-size: 1.2em;
+}
