--- src/components/footer.jsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import './methods.css';
-
-const FunctionalComponents = () => {
-  return (
-    <div className="footer">
-      <span>FAQ</span>
-      <span>Privacy</span>
-      <span>Terms of Use</span>
-      <span>Cookie Preferences</span>
-      <span>Corporate Information</span>
-      <span>Media Center</span>
-      <span>Help Center</span>
-      <span>Contact Us</span>
-    </div>
-  );
-};
-
-export default FunctionalComponents;
