--- src/components/methods.css
@@ -1,64 +1,0 @@
-.logo {
-    background-color: rgb(248, 248, 248);
-    margin: 0;
-    border: rgb(20, 20, 20);
-    text-align: left;
-}
-
-.logo p {
-    color: rgb(229, 9, 20);
-    margin: 0;
-    padding: 20px 0 20px 120px;
-    font-size: 1.2em;
-}
-
-.listMovies {
-    background-color: rgb(235, 235, 235);
-}
-
-.searchTitle {
-    color: rgb(102, 102, 102);
-    margin: 0 0 0 120px;
-    padding-top: 20px;   
-    text-align: left;
-    font-size: 1.5em; 
-}
-
-.searchField {
-    margin: 20px 20px 20px 0;
-    padding: 10px 16px;
-    height: 46px;
-    line-height: 1.33;
-    width: 80%;
-    border: 1px solid rgb(180, 180, 180);
-    border-radius: 2px;
-    box-shadow: inset 0 1px 1px rgba(0,0,0,.075);
-    font-size: 1.2em;
-}
-
-.movies img {
-    margin: 0;
-    padding: 20px;
-    height: 280px;
-    width: 200px;
-}
-
-.footer {
-    margin: 0;
-    padding: 20px 0 20px 100px;
-    text-align: left;
-}
-
-.footer span {
-    color: rgb(102, 102, 102);
-    padding: 20px;
-    width: 25%;
-    border-bottom: 1px solid rgb(255, 255, 255);
-    font-size: 18px;
-}
-
-.legalFooter {
-    color: rgb(102, 102, 102);
-    margin: 0;
-    padding: 20px;
-}
